api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
redis==5.0.8
rq==1.16.2
opencv-python-headless==4.10.0.84
numpy==1.26.4
torch==2.3.0
torchvision==0.18.0
sqlalchemy==2.0.32
alembic==1.13.2
psycopg2-binary==2.9.9
Pillow==10.4.0
